+++ sdk/CHANGELOG.md
@@ -16,7 +16,7 @@ This changelog summarizes major changes between Graal SDK versions. The main foc
 	* MIME types are automatically detected if the source is constructed from a `File` or `URL` if it is not specified explicitly. 
 	* Deprecated `Source.getInputStream()`. Use `Source.getCharacters()` or `Source.getBytes()` instead.
 * Context methods now consistently throw `IllegalArgumentException` instead of `IllegalStateException` for unsupported sources or missing / inaccessible languages.
-
+* Added `Engine.findHome()` to find the GraalVM home folder.
 
 ## Version 1.0 RC5
 * `PolyglotException.getGuestObject()` now returns `null` to indicate that no exception object is available instead of returning a `Value` instance that returns `true` for `isNull()`.